Problem: calculateReward in [/bounties] runs a lot and everywhere
Solution: In bounties.js split away calculateReward to operate on LocalBounties local collection with this.ids.get() filter, and adjust bounties helper query to sort by calculatedReward value of document. Ideally calculateReward would be set to 0 in fields projection in helper that renders the list in bounty.js so that list doesn't rerender unless the sort order changes, and bountyRender having a separate helper that queries for that field to show it in that template. #791