Problem: Blockrazor is not multilingual (#1818)
Solution: Implement TAPi18n translation engine, add a language switcher in the header, and add translation keys instead of strings on the home page, landing page and shared components (header, footer, ...) as an example for the rest of the translation process. Also, translate those pages to Serbian for easier translation testing. New languages will be added to the language switcher automatically. Save user's language choice in a cookie.